Edited to coding standards and quality
