+++ src/main/java/luke/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package luke;
 
 /**
  * This is the parent class to the extended task classes.
@@ -9,6 +9,7 @@ public abstract class Task {
 
     /**
      * Initialises the task class
+     *
      * @param description Description of the task
      */
     public Task(String description) {
@@ -21,28 +22,8 @@ public abstract class Task {
     }
 
     /**
-     * Checks if the task is completed.
-     * @return true is task is completed.
-     */
-    public boolean isDone() {
-        return isDone;
-    }
-
-    /**
-     * Formats the task to be written into the file.
-     * This enables more convenient reading of the file.
-     * @return String of the formatted task
-     */
-    public String textFormatting() {
-        return String.format(this.getClass().getName() + ";" + isDone + ";" + description);
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    /**
      * Gets the X icon to be shown if the task is completed.
+     *
      * @return String of status icon.
      */
     public String getStatusIcon() {
@@ -50,7 +31,32 @@ public abstract class Task {
     }
 
     /**
+     * Checks if the task is completed.
+     *
+     * @return true is task is completed.
+     */
+    public boolean isDone() {
+        return isDone;
+    }
+
+    public void markAsDone() {
+        isDone = true;
+    }
+
+
+    /**
+     * Formats the task to be written into the file.
+     * This enables more convenient reading of the file.
+     *
+     * @return String of the formatted task
+     */
+    public String textFileFormatting() {
+        return String.format(this.getClass().getName() + ";" + isDone + ";" + description);
+    }
+
+    /**
      * Returns a formatted task to the user on the command line interface.
+     *
      * @return String of the formatted task
      */
     public String toString() {
