+++ src/main/java/luke/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package luke;
 
 import java.util.ArrayList;
 
@@ -11,31 +11,33 @@ public class TaskList {
 
     /**
      * Initialises the task list.
+     *
      * @param tasks the list of tasks
      */
     public TaskList(ArrayList<Task> tasks) {
         this.tasks = tasks;
     }
 
-    public Task getTask(int taskNo) {
-        return tasks.get(taskNo);
+    public Task getTask(int taskIndex) {
+        return tasks.get(taskIndex);
     }
 
     /**
      * Gets a task and returns a string with a format to write into the luke file.
      * This format enables easier reading and processing of the tasks in the file.
+     *
      * @param taskIndex Index of the task to format
      * @return String of the formatted task
      */
-    public String getTaskWithFormat(int taskIndex){
-        return tasks.get(taskIndex).textFormatting();
+    public String getTaskWithFileFormat(int taskIndex) {
+        return tasks.get(taskIndex).textFileFormatting();
     }
 
     public int getSize() {
         return tasks.size();
     }
 
-    public void addNewTask(Task addedTask){
+    public void addTask(Task addedTask) {
         tasks.add(addedTask);
     }
 
@@ -43,19 +45,24 @@ public class TaskList {
         tasks.remove(deletedTaskIndex);
     }
 
-    public TaskList findTask(String searchQuery){
+    /**
+     * Returns a new taskList of the tasks that contains the searchQuery.
+     *
+     * @param searchQuery Word or phrase input to search for
+     * @return TaskList of tasks containing the searchQuery
+     */
+    public TaskList findTask(String searchQuery) {
         ArrayList<Task> emptyResults = new ArrayList<>();
         TaskList results = new TaskList(emptyResults);
-        for(Task task : tasks){
-            if(task.getDescription().contains(searchQuery)){
-                results.addNewTask(task);
+        for (Task task : tasks) {
+            if (task.getDescription().contains(searchQuery)) {
+                results.addTask(task);
             }
         }
         return results;
     }
 
     public void checkOffTask(int taskNoCompletedIndex) {
-
         tasks.get(taskNoCompletedIndex).markAsDone();
     }
 }
