+++ src/main/java/luke/commands/AddTask.java
@@ -1,6 +1,6 @@
-package duke.commands;
+package luke.commands;
 
-import duke.Task;
+import luke.Task;
 
 import java.io.IOException;
 
@@ -8,7 +8,7 @@ import java.io.IOException;
  * This class adds a tasks to the task list.
  */
 public class AddTask extends Command {
-    private Task task;
+    private final Task task;
 
     public static final String TODO_COMMAND = "todo";
     public static final String EVENT_COMMAND = "event";
@@ -16,22 +16,24 @@ public class AddTask extends Command {
 
     /**
      * Initialises the AddTask class.
+     *
      * @param task Task that will be added.
      */
 
-    public AddTask(Task task){
+    public AddTask(Task task) {
         this.task = task;
     }
 
     /**
-     * Executes the adding the task to the task list.
-     * It shows the AddTask messages through the ui.
-     * It saves the data in the file.
+     * Executes adding the task to the task list.
+     * Shows the AddTask messages through the ui.
+     * Saves the data in the file.
+     *
      * @throws IOException Thrown when there is an error writing to the file.
      */
     @Override
     public void execute() throws IOException {
-        tasks.addNewTask(task);
+        tasks.addTask(task);
         ui.showTaskAddedMessage(task, tasks);
         storage.writeFile(tasks);
     }
