+++ src/main/java/luke/commands/CheckOffTask.java
@@ -1,18 +1,19 @@
-package duke.commands;
+package luke.commands;
 
-import duke.customexception.IllegalNumberInputException;
+import luke.customexception.IllegalNumberInputException;
 
 import java.io.IOException;
 
 /**
  * This class checks off a task in the task list as done.
  */
-public class CheckOffTask extends Command{
-    private int taskNoCompletedIndex;
+public class CheckOffTask extends Command {
+    private final int taskNoCompletedIndex;
     public static final String COMMAND_WORD = "done";
 
     /**
      * Initialises the CheckOffTask class.
+     *
      * @param taskNoCompletedIndex Index of the task that is completed.
      */
     public CheckOffTask(int taskNoCompletedIndex) {
@@ -23,18 +24,18 @@ public class CheckOffTask extends Command{
      * Executes checking off the task in the task list.
      * Shows the message after checking off the task.
      * Stores the data by writing into the file
-     * @throws IOException Thrown when there is an error writing into the file
+     *
+     * @throws IOException                 Thrown when there is an error writing into the file
      * @throws IllegalNumberInputException Thrown when the index provided is not within the size of the task list.
      */
     @Override
     public void execute() throws IOException, IllegalNumberInputException {
-        if((taskNoCompletedIndex < 0) || (taskNoCompletedIndex >= tasks.getSize())){
+        if ((taskNoCompletedIndex < 0) || (taskNoCompletedIndex >= tasks.getSize())) {
             throw new IllegalNumberInputException();
         }
-        if(tasks.getTask(taskNoCompletedIndex).isDone()){
+        if (tasks.getTask(taskNoCompletedIndex).isDone()) {
             ui.showTaskCompletedAlreadyMessage(tasks.getTask(taskNoCompletedIndex));
-        }
-        else {
+        } else {
             tasks.checkOffTask(taskNoCompletedIndex);
             ui.showTaskCompletedMessage(tasks.getTask(taskNoCompletedIndex));
         }
