+++ src/main/java/luke/commands/DeleteTask.java
@@ -1,19 +1,20 @@
-package duke.commands;
+package luke.commands;
 
-import duke.customexception.IllegalNumberInputException;
+import luke.customexception.IllegalNumberInputException;
 
 import java.io.IOException;
 
 /**
  * This class deletes a task from the task list.
  */
-public class DeleteTask extends Command{
+public class DeleteTask extends Command {
 
-    private int deletedTaskIndex;
+    private final int deletedTaskIndex;
     public static final String COMMAND_WORD = "delete";
 
     /**
      * Initialises the DeleteTask class.
+     *
      * @param deletedTaskIndex Index of the task that will be deleted.
      */
     public DeleteTask(int deletedTaskIndex) {
@@ -24,11 +25,12 @@ public class DeleteTask extends Command{
      * Executes deleting the task from the task list.
      * Shows delete task messages.
      * Stores the information by writing into the file.
-     * @throws IOException
+     *
+     * @throws IOException Thrown when there is an error writing to file
      */
     @Override
     public void execute() throws IOException, IllegalNumberInputException {
-        if(deletedTaskIndex < 0 || deletedTaskIndex >= tasks.getSize()){
+        if (deletedTaskIndex < 0 || deletedTaskIndex >= tasks.getSize()) {
             throw new IllegalNumberInputException();
         }
         ui.showTaskDeletedMessage(tasks.getTask(deletedTaskIndex), tasks);
