+++ src/main/java/luke/customexception/EmptyNumberInputException.java
@@ -1,13 +1,13 @@
-package duke.customexception;
+package luke.customexception;
 
-import static duke.common.Messages.FUNCTIONS_USER_GUIDE_MESSAGE;
+import static luke.common.Messages.FUNCTIONS_USER_GUIDE_MESSAGE;
 
 /**
  * This class is thrown when there is no number input when it is expected to have one.
  */
-public class EmptyNumberInputException extends Exception{
+public class EmptyNumberInputException extends Exception {
     @Override
-    public String getMessage(){
+    public String getMessage() {
         String errorMessage = "Please input a valid command number!\n" + FUNCTIONS_USER_GUIDE_MESSAGE;
         return errorMessage;
     }
