+++ src/main/java/luke/customexception/EmptyTaskInputException.java
@@ -1,13 +1,13 @@
-package duke.customexception;
+package luke.customexception;
 
-import static duke.common.Messages.TASKS_USER_GUIDE_MESSAGE;
+import static luke.common.Messages.TASKS_USER_GUIDE_MESSAGE;
 
 /**
  * This class is thrown when the description of the task is empty.
  */
-public class EmptyTaskInputException extends Exception{
+public class EmptyTaskInputException extends Exception {
     @Override
-    public String getMessage(){
+    public String getMessage() {
         String errorMessage = "Invalid task format!\n"
                 + TASKS_USER_GUIDE_MESSAGE;
         return errorMessage;
