+++ src/main/java/luke/customexception/EmptyTimeException.java
@@ -1,13 +1,13 @@
-package duke.customexception;
+package luke.customexception;
 
-import static duke.common.Messages.TASKS_USER_GUIDE_MESSAGE;
+import static luke.common.Messages.TASKS_USER_GUIDE_MESSAGE;
 
 /**
  * This class is thrown when the deadline date or event duration is empty.
  */
 public class EmptyTimeException extends Exception {
     @Override
-    public String getMessage(){
+    public String getMessage() {
         String errorMessage = "Invalid event duration or deadline date input\n"
                 + TASKS_USER_GUIDE_MESSAGE;
         return errorMessage;
