+++ src/main/java/luke/customexception/IllegalFindInputException.java
@@ -1,8 +1,11 @@
-package duke.customexception;
+package luke.customexception;
 
-public class IllegalFindInputException extends Exception{
+/**
+ * This class is thrown when the searchQuery is empty
+ */
+public class IllegalFindInputException extends Exception {
     @Override
-    public String getMessage(){
+    public String getMessage() {
         String errorMessage = "Invalid search input!\n"
                 + "find: Finds a keyword and lists all relevant tasks"
                 + " PARAMETERS: SEARCH_QUERY\n"
