+++ src/main/java/luke/tasktype/Deadline.java
@@ -1,14 +1,13 @@
-package duke.tasktype;
+package luke.tasktype;
 
-import duke.Task;
+import luke.Task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 /**
  * This class stores the information of a deadline task
- * It has an additional deadlineDate trait which stores
- * when this task is due in a YYYY-MM-DD format.
+ * DeadlineDate trait stores when task is due in a YYYY-MM-DD format.
  */
 
 public class Deadline extends Task {
@@ -17,7 +16,8 @@ public class Deadline extends Task {
 
     /**
      * Initialises the deadline class
-     * @param description Description of the deadline
+     *
+     * @param description  Description of the deadline
      * @param deadlineDate The deadline date
      */
     public Deadline(String description, LocalDate deadlineDate) {
@@ -32,15 +32,17 @@ public class Deadline extends Task {
 
     /**
      * Returns a string of the formatted deadline to be written into the file
+     *
      * @return String of the formatted deadline
      */
     @Override
-    public String textFormatting() {
-        return String.format(super.textFormatting() + ";" + deadlineDate);
+    public String textFileFormatting() {
+        return String.format(super.textFileFormatting() + ";" + deadlineDate);
     }
 
     /**
      * Returns the string of the formatted deadline to the command line interface
+     *
      * @return String of the formatted deadline.
      */
     @Override
